Move history fetching to transport app
The goal is to make stubbing easier
